[Refactoring ⚙️] Bottom 공통 컴포넌트 추상화 제거
