+++ .prettierrc
@@ -10,6 +10,7 @@
     "^@apis/(.*)$",
     "^@assets/(.*)$",
     "^@pages/(.*)$",
+    "^@common/(.*)$",
     "^@typpes/(.*)$",
     "^@routes/(.*)$",
     "^@hooks/(.*)$",
